pyproject.toml
[project]
name = "project05"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fal-client>=0.13.2",
    "gradio>=6.11.0",
    "mcp[cli]>=1.26.0",
    "openai>=2.30.0",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
]
